disabled change_set.cc and all dependencies, got all of roster4 building

+++ file_io.cc
@@ -130,7 +130,7 @@ static bool char_is_binary[256];
 void
 set_char_is_binary(char c, bool is_binary)
 {
-    char_is_binary[c] = is_binary;
+  char_is_binary[static_cast<size_t>(c)] = is_binary;
 }
 
 static void
@@ -144,9 +144,9 @@ init_char_is_binary()
 		       "\x10\x11\x12\x13\x14\x15\x16\x17\x18"
 		       "\x19\x1a\x1c\x1d\x1e\x1f");
   set_char_is_binary('\0',true);
-  for(int i=0;i<nontext_chars.size();++i) 
+  for(size_t i = 0; i < nontext_chars.size(); ++i) 
     {
-      set_char_is_binary(nontext_chars[i],true);
+      set_char_is_binary(nontext_chars[i], true);
     }
 }
 
@@ -157,9 +157,9 @@ bool guess_binary(string const & s)
       init_char_is_binary();
     }
 
-  for(int i = 0;i<s.size(); ++i) 
+  for(size_t i = 0; i < s.size(); ++i) 
     {
-      if (char_is_binary[s[i]])
+      if (char_is_binary[static_cast<size_t>(s[i])])
         return true;
     }
   return false;
